phone int

+++ php/feedback.php
@@ -41,15 +41,15 @@
             echo 'Спасибо, ' .$name. '. Ваша заявка отправлена. <br>';
 
 
-            $phone_repl = str_replace('+7', '', $phone);
+            $phone_repl = (int)str_replace('+7', '', $phone);
             echo 'Номер телефона в базу ' . $phone_repl . ' было ' . $phone;
             $now_time = time();
             var_dump($phone_repl);
 
             if ($textarea != NULL) {
-                $request = "INSERT INTO `orders`(`name`, `phone`, `message`, `id_form`, `time`) VALUES ('".$name."',".(int)$phone_repl.",'".$textarea."', ".$id_form.", ". $now_time .")";
+                $request = "INSERT INTO `orders`(`name`, `phone`, `message`, `id_form`, `time`) VALUES ('".$name."',".$phone_repl.",'".$textarea."', ".$id_form.", ". $now_time .")";
             } else {
-                $request = "INSERT INTO `orders`(`name`, `phone`, `message`, `id_form`, `time`) VALUES ('".$name."',".(int)$phone_repl.",NULL, ".$id_form.", ". $now_time .")";
+                $request = "INSERT INTO `orders`(`name`, `phone`, `message`, `id_form`, `time`) VALUES ('".$name."',".$phone_repl.",NULL, ".$id_form.", ". $now_time .")";
             }
             // echo '<br>' . $request;
             $result_mysql = $connection->exec($request);
